refactoring .. wip ..
git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/branches/JOGL_2_SANDBOX@1656 232f8b59-042b-4e1e-8c03-345bb8c30851

+++ src/classes/com/sun/javafx/newt/NewtFactory.java
@@ -72,7 +72,7 @@ public abstract class NewtFactory {
       return Screen.create(getWindowType(), display, index);
     }
 
-    public static Window createWindow(Screen screen, long visualID) {
+    public static Window createWindow(Screen screen, int visualID) {
       return Window.create(getWindowType(), screen, visualID);
     }
